requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
python-dateutil>=2.8.2
Pillow>=9.0.0
psycopg2-binary>=2.9.5
faker>=13.0.0
django-filter>=23.1
django-cors-headers>=4.0.0